Time: 4 ms (67.05%), Space: 6.9 MB (89.15%) - LeetHub

+++ 8-string-to-integer-atoi/8-string-to-integer-atoi.cpp
@@ -30,6 +30,6 @@ public:
             
             i++;
         }
-        return (int) (sign * res);
+        return (sign * res);
     }
 };
